Move profiles into dedicated submodules.
