+++ src/fixture/profile/comet.rs
@@ -4,7 +4,7 @@ use number::UnipolarFloat;
 use std::{collections::VecDeque, time::Duration};
 
 use super::generic::{GenericStrobe, GenericStrobeStateChange};
-use super::prelude::*;
+use crate::fixture::prelude::*;
 use crate::util::unipolar_to_range;
 
 #[derive(Default, Debug)]
