+++ src/fixture/profile/hypnotic.rs
@@ -3,8 +3,8 @@
 use anyhow::Context;
 use number::BipolarFloat;
 
-use super::animation_target::TargetedAnimationValues;
-use super::prelude::*;
+
+use crate::fixture::prelude::*;
 use crate::util::bipolar_to_split_range;
 use num_derive::{FromPrimitive, ToPrimitive};
