+++ src/fixture/profile/mod.rs
@@ -1,0 +1,21 @@
+pub mod aquarius;
+pub mod astroscan;
+pub mod color;
+pub mod colordynamic;
+pub mod comet;
+pub mod dimmer;
+pub mod faderboard;
+pub mod freedom_fries;
+pub mod generic;
+pub mod h2o;
+pub mod hypnotic;
+pub mod lumasphere;
+pub mod radiance;
+pub mod rotosphere_q3;
+pub mod rush_wizard;
+pub mod solar_system;
+pub mod starlight;
+pub mod swarmolon;
+pub mod uv_led_brick;
+pub mod venus;
+pub mod wizard_extreme;
