+++ src/fixture/profile/solar_system.rs
@@ -5,8 +5,7 @@ use log::error;
 use num_derive::{FromPrimitive, ToPrimitive};
 use number::BipolarFloat;
 
-use super::animation_target::TargetedAnimationValues;
-use super::prelude::*;
+use crate::fixture::prelude::*;
 use crate::util::unipolar_to_range;
 use strum_macros::{Display as EnumDisplay, EnumIter, EnumString};
