+++ src/fixture/profile/venus.rs
@@ -5,7 +5,7 @@ use std::time::Duration;
 use anyhow::Context;
 use number::{BipolarFloat, UnipolarFloat};
 
-use super::prelude::*;
+use crate::fixture::prelude::*;
 use crate::util::{unipolar_to_range, RampingParameter};
 
 /// Control abstraction for the RA venus.
