+++ src/osc/mod.rs
@@ -23,9 +23,9 @@ use self::radio_button::{EnumRadioButton, RadioButton};
 mod animation;
 mod basic_controls;
 mod control_message;
-mod fixture;
 mod label_array;
 mod master;
+mod profile;
 mod radio_button;
 mod register;
 mod show;
